Resolve bugprone-implicit-widening-of-multiplication-result ClangTidy warnings on my local build

+++ core/src/Threads/Kokkos_Threads_Team.hpp
@@ -766,7 +766,7 @@ class TeamPolicyInternal<Kokkos::Threads, Properties...>
  private:
   /** \brief finalize chunk_size if it was set to AUTO*/
   inline void set_auto_chunk_size() {
-    int64_t concurrency = space().concurrency() / m_team_alloc;
+    int concurrency = space().concurrency() / m_team_alloc;
     if (concurrency == 0) concurrency = 1;
 
     if (m_chunk_size > 0) {
